Use updated yaru_widgets
- this needs an update to how the search works
- took the opportunity and move the app into its own file
